restructured modal file system and created a socialicons option for sharing
