+++ client/src/components/widgets/modals-and-materials/OptionIcons.jsx
@@ -1,0 +1,38 @@
+import React from 'react'
+
+export default function OptionIcons(props) {
+  return (
+    <>
+      {/* EDIT */}
+      <div className='option-icon-container'>
+        <svg width="46" height="58" viewBox="0 0 46 58" fill="none" xmlns="http://www.w3.org/2000/svg">
+          <g clipPath="url(#clip0)">
+            <path d="M42.9449 15.0294L35.9143 8.02345C35.567 7.67732 35.0023 7.67732 34.655 8.02345L7.68327 34.8969C7.67614 34.9041 7.67258 34.9147 7.66545 34.9236C7.64586 34.9449 7.63517 34.9697 7.61914 34.9946C7.54611 35.0904 7.49445 35.1934 7.46595 35.3052C7.46061 35.3265 7.44102 35.3407 7.43745 35.3602L5.80939 43.9849C5.75595 44.2725 5.8468 44.5689 6.0552 44.7766C6.22264 44.9452 6.44886 45.0375 6.68398 45.0375C6.73742 45.0375 6.79442 45.0322 6.84786 45.0215L15.5083 43.4063C15.5315 43.4027 15.5457 43.3832 15.5653 43.3779C15.6793 43.346 15.7844 43.2962 15.8806 43.2235C15.9037 43.2057 15.9269 43.1968 15.9483 43.1791C15.9572 43.172 15.9679 43.1684 15.975 43.1614L42.9467 16.2825C43.294 15.9364 43.294 15.3755 42.9449 15.0294ZM41.0568 15.6559L38.8534 17.8516L33.0821 12.1006L35.2855 9.90495L41.0568 15.6559ZM31.821 13.3555L37.5922 19.1065L15.3444 41.2798L12.4588 38.4026L9.57139 35.5253L31.821 13.3555ZM7.80083 43.0389L8.8767 37.3411L11.1977 39.6539L13.5222 41.9721L7.80083 43.0389Z" fill="black" />
+          </g>
+          <defs>
+            <clipPath id="clip0">
+              <rect width="46" height="58" fill="white" />
+            </clipPath>
+          </defs>
+        </svg>
+        <h3 className='option-label'>Edit</h3>
+      </div>
+      {/* DELETE */}
+      <div className='option-icon-container'>
+        <svg width="46" height="58" viewBox="0 0 46 58" fill="none" xmlns="http://www.w3.org/2000/svg">
+          <path d="M21.4187 4.32825C19.7074 4.32825 18.3036 5.63995 18.2131 7.31525H12.88C10.8571 7.31525 9.19995 8.98685 9.19995 11.0272V11.0562C9.19995 12.2764 9.81466 13.3655 10.7237 14.0432C10.0732 14.7137 9.65995 15.6331 9.65995 16.6387V38.3597C9.65995 40.4001 11.3171 42.0717 13.34 42.0717H32.6887C34.7115 42.0717 36.3687 40.4001 36.3687 38.3597V16.6387C36.3687 15.6258 35.9638 14.7005 35.305 14.0287C36.2101 13.3506 36.8 12.2733 36.8 11.0562V11.0272C36.8 8.98685 35.1428 7.31525 33.12 7.31525H27.7868C27.6963 5.63995 26.2925 4.32825 24.5812 4.32825H21.4187ZM21.4187 5.25625H24.5812C25.8097 5.25625 26.7779 6.1542 26.8668 7.31525H19.1331C19.222 6.1542 20.1902 5.25625 21.4187 5.25625ZM12.88 9.17125H33.12C34.1552 9.17125 34.96 9.98297 34.96 11.0272V11.0562C34.96 12.1005 34.1552 12.9122 33.12 12.9122H12.88C11.8447 12.9122 11.04 12.1005 11.04 11.0562V11.0272C11.04 9.98297 11.8447 9.17125 12.88 9.17125ZM13.34 14.7827H16.1575V40.2157H13.34C12.3047 40.2157 11.5 39.404 11.5 38.3597V16.6387C11.5 15.5945 12.3047 14.7827 13.34 14.7827ZM17.02 14.7827H22.5975V40.2157H17.02V14.7827ZM23.4312 14.7827H29.0375V40.2157H23.4312V14.7827ZM29.8712 14.7827H32.6887C33.724 14.7827 34.5287 15.5945 34.5287 16.6387V38.3597C34.5287 39.404 33.724 40.2157 32.6887 40.2157H29.8712V14.7827Z" fill="black" fillOpacity="0.980392" />
+        </svg>
+        <h3 className='option-label'>Delete</h3>
+      </div>
+      {/* SHARE */}
+      <div className='option-icon-container' onClick={props.showIcons}>
+        <svg width="71" height="71" viewBox="0 0 71 71" fill="none" xmlns="http://www.w3.org/2000/svg">
+          <path d="M43.026 19.454L35.5 11.928L27.974 19.454L25.986 17.466L35.5 7.95203L45.014 17.466L43.026 19.454Z" fill="black" />
+          <path d="M34.08 9.93994H36.92V39.7599H34.08V9.93994Z" fill="black" />
+          <path d="M49.7 56.8H21.3C18.886 56.8 17.04 54.954 17.04 52.54V26.98C17.04 24.566 18.886 22.72 21.3 22.72H31.24V25.56H21.3C20.448 25.56 19.88 26.128 19.88 26.98V52.54C19.88 53.392 20.448 53.96 21.3 53.96H49.7C50.552 53.96 51.12 53.392 51.12 52.54V26.98C51.12 26.128 50.552 25.56 49.7 25.56H39.76V22.72H49.7C52.114 22.72 53.96 24.566 53.96 26.98V52.54C53.96 54.954 52.114 56.8 49.7 56.8Z" fill="black" />
+        </svg>
+        <h3 className='option-label'>Share</h3>
+      </div>
+    </>
+  )
+}
