+++ client/src/components/widgets/modals-and-materials/SocialShareIcons.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function SocialShareModal() {
+  return (
+    <>
+      YOoo
+    </>
+  )
+}
